+++ src/cpp/session/SessionMain.cpp
@@ -655,7 +655,9 @@ void handleClientInit(const boost::function<void()>& initFunction,
 
    sessionInfo["clang_available"] = modules::clang::isAvailable();
 
-   sessionInfo["show_help_home"] = options.showHelpHome();
+   // don't show help home until we figure out a sensible heuristic
+   // sessionInfo["show_help_home"] = options.showHelpHome();
+   sessionInfo["show_help_home"] = false;
 
    // send response  (we always set kEventsPending to false so that the client
    // won't poll for events until it is ready)