+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -2553,6 +2553,11 @@ public class TextEditingTarget implements
                   0));
    }
    
+   public void showHelpAtCursor()
+   {
+      docDisplay_.goToHelp();
+   }
+   
    @Handler
    void onDebugBreakpoint()
    {